Refine audio visualizer config narrowing
